requirements.txt
asgiref==3.3.1
certifi==2020.6.20
Django==3.1.3
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
wincertstore==0.2